move connection process to network.py
